+++ src/gevent/libev/_corecffi_build.py
@@ -61,7 +61,7 @@ void vfd_free(int);
 
 include_dirs = [
     thisdir, # libev_vfd.h
-    os.path.abspath(os.path.join(thisdir, '..', '..', 'deps', 'libev')),
+    os.path.abspath(os.path.join(thisdir, '..', '..', '..', 'deps', 'libev')),
 ]
 ffi.cdef(_cdef)
 ffi.set_source('gevent._corecffi', _source, include_dirs=include_dirs)