+++ src/gevent/_ssl3.py
@@ -221,10 +221,11 @@ class SSLSocket(socket):
         Return zero-length string on EOF."""
         # pylint:disable=too-many-branches
         self._checkClosed()
-        if not self._sslobj:
-            raise ValueError("Read on closed or unwrapped SSL socket.")
 
         while True:
+            if not self._sslobj:
+                raise ValueError("Read on closed or unwrapped SSL socket.")
+
             try:
                 if buffer is not None:
                     return self._sslobj.read(len, buffer)
@@ -252,10 +253,11 @@ class SSLSocket(socket):
         """Write DATA to the underlying SSL channel.  Returns
         number of bytes of DATA actually transmitted."""
         self._checkClosed()
-        if not self._sslobj:
-            raise ValueError("Write on closed or unwrapped SSL socket.")
 
         while True:
+            if not self._sslobj:
+                raise ValueError("Write on closed or unwrapped SSL socket.")
+
             try:
                 return self._sslobj.write(data)
             except SSLError as ex: